enemyskillprison만 추가했습니다. 나머지 스킬은 나중에 추가하겠습니다

+++ TeamProjectFinal/SW설계기초-Final Project/UI.h
@@ -79,9 +79,9 @@ void ShowMap() {
 void InvalidateMap() {
 	COORD ptr = { BACKGROUND_ORIGIN_X + BACKGROUND_ROW, BACKGROUND_ORIGIN_Y };
 	int modifiedX = ptr.X - 32, modifiedY = ptr.Y + 1;
-	ShowPlayerPosCursor(modifiedX + CurrentTime - 1, modifiedY + 1);
+	ShowPlayerPosCursor(modifiedX + (int)(CurrentTime) - 1, modifiedY + 1);
 	if (TimeCheckerEnd() - PausingTime > CurrentTime * (StageTime[StageNumber - 1] / (MAPLENGTH * 2))) {
-		HidePlayerPosCursor(modifiedX + CurrentTime - 1, modifiedY + 1);
+		HidePlayerPosCursor(modifiedX + (int)(CurrentTime) - 1, modifiedY + 1);
 		CurrentTime++; PlayerPos++;
 	}
 	SetCurrentCursorPos(GAMEBOARD_ORIGIN_X, GAMEBOARD_ORIGIN_Y);
